+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNS4KN2ZJK9890Z23Y0P7SMW"
+        "01KNSDDF1RSSS2JJWXX6R80BBD"
     )
     r"""The unique identifier of the data source"""