+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNS4KN2Y22X5SY233H6MV41R"
+        "01KNSDDF1QYNCA4WT96A27YAN6"
     )
     r"""The unique identifier of the data source"""