+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNS4KN31QVYVECZ0K1Y462JX"
+        "01KNSDDF1T3D8YXWVP5XBQW48M"
     )
     r"""The unique identifier of the data source"""