+++ packages/orq-rc/RELEASES.md
@@ -7009,3 +7009,13 @@ Based on:
 - [python v4.8.0-rc.0] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.0 - packages/orq-rc
+
+## 2026-04-09 15:26:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.1 - packages/orq-rc